refactor: replaced StationDetailPage for a modal in mobile
